batch: refactoring the validators out
