+++ azurerm/internal/services/batch/batch_certificate_data_source.go
@@ -2,8 +2,9 @@ package batch
 
 import (
 	"fmt"
-	"regexp"
 	"time"
+
+	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/batch/validate"
 
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/helpers/azure"
@@ -24,13 +25,13 @@ func dataSourceBatchCertificate() *schema.Resource {
 			"name": {
 				Type:         schema.TypeString,
 				Required:     true,
-				ValidateFunc: validateAzureRMBatchCertificateName,
+				ValidateFunc: validate.CertificateName,
 			},
 
 			"account_name": {
 				Type:         schema.TypeString,
 				Required:     true,
-				ValidateFunc: ValidateAzureRMBatchAccountName,
+				ValidateFunc: validate.AccountName,
 			},
 
 			"resource_group_name": azure.SchemaResourceGroupNameForDataSource(),
@@ -86,14 +87,4 @@ func dataSourceBatchCertificateRead(d *schema.ResourceData, meta interface{}) er
 	}
 
 	return nil
-}
-
-func validateAzureRMBatchCertificateName(v interface{}, k string) (warnings []string, errors []error) {
-	value := v.(string)
-	if !regexp.MustCompile(`^[\w]+-[\w]+$`).MatchString(value) {
-		errors = append(errors, fmt.Errorf(
-			"must be made up of algorithm and thumbprint separated by a dash in %q: %q", k, value))
-	}
-
-	return warnings, errors
 }
